feat: Replace SmartLockUI with FrontDoorUI, add package pickup notification
- Renamed SmartLockUI to FrontDoorUI to better reflect its purpose.
- Commented out the DigitalLease route since the file has not been added.
- Implemented a toast notification in the tenant's smart lock view to inform tenant they have successfully picked up a package.
